Refactor the test XML utilities
