+++ tst/xmlsamples/CMakeLists.txt
@@ -2,4 +2,4 @@ configure_file(settings-fragment.xml . COPYONLY)
 configure_file(settings-comment-fragment.xml . COPYONLY)
 configure_file(devicerequest-binaryinput.xml . COPYONLY)
 configure_file(devicerequest-binaryoutput.xml . COPYONLY)
-configure_file(devicerequiest-pwmchannel.xml . COPYONLY)
+configure_file(devicerequest-pwmchannel.xml . COPYONLY)
